mixed: move linear operators to separate file
